pyproject.toml
[tool.poetry]
name = "codeforces-parser"
version = "0.1.0"
description = ""
authors = ["RussianProgram <kurban.keirmhanov@gmail.com>"]
readme = "README.md"
packages = [{include = "codeforces_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
psycopg-binary = "^3.1.8"
psycopg2 = "^2.9.5"
python-telegram-bot = "^20.1"
aiogram = "2.12.1"
sqlalchemy = "1.4.23"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
